+++ src/components/DimensionsPanel/styles/DimensionsPanel.style.js
@@ -6,10 +6,9 @@ export const styles = {
         display: 'flex',
         flexDirection: 'column',
         backgroundColor: colors.snow,
-        padding: '8px 0 2px 8px',
+        padding: '8px',
     },
     textField: {
-        paddingBottom: '12px',
-        width: 'calc(100% - 8px)',
+        paddingBottom: '6px',
     },
 }