+++ src/visualizations/config/adapters/dhis_dhis/index.js
@@ -2,7 +2,7 @@ import getTitle from './title'
 import getSubtitle from './subtitle'
 
 export default function({ store, layout, extraOptions }) {
-    const value = store.generateData({
+    const data = store.generateData({
         type: layout.type,
         seriesId:
             layout.columns && layout.columns.length
@@ -12,8 +12,10 @@ export default function({ store, layout, extraOptions }) {
             layout.rows && layout.rows.length ? layout.rows[0].dimension : null,
     })
 
+    const value = data[0] === undefined ? extraOptions.noData.text : data[0]
+
     const config = {
-        value: value[0],
+        value: value,
         title: getTitle(layout, store.data[0].metaData, extraOptions.dashboard),
         subtitle: getSubtitle(
             layout,