+++ src/visualizations/config/adapters/dhis_dhis/subtitle/singleValue.js
@@ -1,0 +1,5 @@
+import getFilterText from '../getFilterText'
+
+export default function(layout, metaData) {
+    return layout.filters ? getFilterText(layout.filters, metaData) : ''
+}